Rename folder core to util
