+++ index.ts
@@ -1,4 +1,5 @@
 import { NgModule } from '@angular/core';
+
 import { OFILEMANAGER_MODULES } from './src/components';
 import { OFILEMANAGER_PROVIDERS } from './src/services';
 
@@ -6,7 +7,7 @@ import { OFILEMANAGER_PROVIDERS } from './src/services';
  * Exports
  */
 export * from './src/components';
-export * from './src/core';
+export * from './src/util';
 export * from './src/services';
 
 @NgModule({
@@ -14,5 +15,4 @@ export * from './src/services';
   exports: OFILEMANAGER_MODULES,
   providers: OFILEMANAGER_PROVIDERS
 })
-export class OFileManagerModule {
-}
+export class OFileManagerModule { }
