+++ src/components/filemanager-table/table-extended/dialog/changename/change-name-dialog.component.ts
@@ -1,8 +1,8 @@
 import { AfterViewInit, Component, ElementRef, Injector, ViewChild } from '@angular/core';
 import { FormControl, Validators } from '@angular/forms';
-import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material';
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material';
 
-import { FileClass, OFileManagerTranslatePipe, fileNameValidator } from '../../../../../core';
+import { FileClass, fileNameValidator, OFileManagerTranslatePipe } from '../../../../../util';
 
 export class ChangeNameDialogData {
   title: string;
