--- src/core.ts
@@ -1,3 +1,0 @@
-export * from './core/file.class';
-export * from './core/o-filemanager-translate.pipe';
-export * from './core/filename.validator';
